+++ packages/icons/src/svg/index.js
@@ -1,19 +1,21 @@
-import Star from './star.svg';
-import Clock from './clock.svg';
+import ArrowDown from './arrow_down.svg';
 import Building from './building.svg';
-import TriangleAlert from './triangle_alert.svg';
+import Clock from './clock.svg';
 import Close from './close.svg';
+import Done from './done.svg';
+import Star from './star.svg';
+import TriangleAlert from './triangle_alert.svg';
 import Visibility from './visibility.svg';
 import VisibilityOff from './visibility_off.svg';
-import ArrowDown from './arrow_down.svg';
 
 export {
-  Star,
-  Clock,
+  ArrowDown,
   Building,
-  TriangleAlert,
+  Clock,
   Close,
+  Done,
+  Star,
+  TriangleAlert,
   Visibility,
   VisibilityOff,
-  ArrowDown,
 };