+++ packages/doc/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.3](https://github.com/Gympass/yoga/compare/@gympass/yoga-doc@0.17.2...@gympass/yoga-doc@0.17.3) (2020-03-12)
+
+**Note:** Version bump only for package @gympass/yoga-doc
+
 ## [0.17.2](https://github.com/Gympass/yoga/compare/@gympass/yoga-doc@0.17.1...@gympass/yoga-doc@0.17.2) (2020-03-10)
 
 **Note:** Version bump only for package @gympass/yoga-doc